Restore legacy stack canary behavior.
The version of NetBSD that Rumprun uses expects the stack canary to be
stored as a global variable `__stack_chk_guard`, while newer versions of
GCC expect it to be in thread-local storage. This resulted in the canary
value in TLS never being initialized, i.e. always being 0.

This commit re-enables the old canary behavior so that the
canary-related code that GCC inserts will look for its value in the
global variable. In addition, it replaces the pervasive uses of
`-fno-stack-protector` throughout the codebase with
`-fstack-protector-strong` to ensure that canaries are used
all over the application.